+++ src/module/applications/sheets/actor-sheet.mjs
@@ -972,22 +972,17 @@ export default class DrawSteelActorSheet extends DSDocumentSheetMixin(sheets.Act
   async _onDropFolder(event, data) {
     if (!this.actor.isOwner) return [];
     const folder = await Folder.implementation.fromDropData(data);
-    if (folder.type !== "Item") return [];
-    const projectDropTarget = event.target.closest("[data-application-part='projects'");
+    if (folder.type !== "Item") return []; // V14 - handle ActiveEffect
     const droppedItemData = await Promise.all(
       folder.contents.map(async (item) => {
         if (!(document instanceof Item)) item = await fromUuid(item.uuid);
 
-        // If it's an equipment dropped on the project tab, create the item as a project
-        if (projectDropTarget && (item.type === "equipment")) {
-          const name = game.i18n.format("DRAW_STEEL.Item.project.Craft.ItemName", { name: item.name });
-          item = { name, type: "project", "system.yield.item": item.uuid };
-        }
+        const keepId = !this.actor.items.has(item.id);
 
-        return item;
+        return game.items.fromCompendium(item, { keepId, clearFolder: true });
       }),
     );
-    this.actor.createEmbeddedDocuments("Item", droppedItemData);
+    this.actor.createEmbeddedDocuments("Item", droppedItemData, { keepId: true });
   }
 
   /* -------------------------------------------------- */